refactor: update to db-writer-common 4.0
