+++ Dockerfile
@@ -1,26 +1,33 @@
-FROM php:5.6
-MAINTAINER Miroslav Cillik <miro@keboola.com>
+FROM php:7-cli
+ARG DEBIAN_FRONTEND=noninteractive
+ARG COMPOSER_FLAGS="--prefer-dist --no-interaction"
+ENV COMPOSER_ALLOW_SUPERUSER=1
+ENV COMPOSER_PROCESS_TIMEOUT 3600
 
 # Deps
 RUN apt-get update
-RUN apt-get install -y wget curl make git bzip2 time libzip-dev openssl
+RUN apt-get install -y wget curl make git bzip2 time libzip-dev openssl unzip
 
 # PHP
 RUN docker-php-ext-install pdo pdo_mysql
-
-WORKDIR /home
+RUN echo "memory_limit = -1" >> /usr/local/etc/php/php.ini
+RUN echo "date.timezone = \"Europe/Prague\"" >> /usr/local/etc/php/php.ini
 
 # Composer
 WORKDIR /root
-RUN cd \
-  && curl -sS https://getcomposer.org/installer | php \
-  && ln -s /root/composer.phar /usr/local/bin/composer
+RUN curl -sS https://getcomposer.org/installer | php \
+  && mv composer.phar /usr/local/bin/composer
 
 # Main
-ADD . /code
 WORKDIR /code
-RUN echo "memory_limit = -1" >> /usr/local/etc/php/php.ini
-RUN echo "date.timezone = \"Europe/Prague\"" >> /usr/local/etc/php/php.ini
-RUN composer selfupdate && composer install --no-interaction
+## Composer - deps always cached unless changed
+# First copy only composer files
+COPY composer.* /code/
+# Download dependencies, but don't run scripts or init autoloaders as the app is missing
+RUN composer install $COMPOSER_FLAGS --no-scripts --no-autoloader
+# copy rest of the app
+COPY . /code/
+# run normal composer - all deps are cached already
+RUN composer install $COMPOSER_FLAGS
 
 CMD php ./run.php --data=/data
