+++ composer.json
@@ -11,17 +11,43 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "keboola/db-writer-common": "2.2.*",
-        "keboola/php-temp": "^0.1.3"
+        "php": "^7.1",
+        "keboola/db-writer-common": "^4.0",
+        "keboola/php-temp": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.6.*"
+        "keboola/coding-standard": "^2.0",
+        "phpstan/phpstan-shim": "^0.9.2",
+        "phpunit/phpunit": "^7.0"
     },
     "autoload": {
         "psr-4": {
-            "Keboola\\": "src/Keboola/"
+            "Keboola\\DbWriter\\": "src/"
         }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Keboola\\DbWriter\\Tests\\": "tests/phpunit/"
+        }
+    },
+    "scripts": {
+        "tests": "phpunit",
+        "phpcs": "phpcs -n --ignore=vendor --extensions=php .",
+        "phpcbf": "phpcbf -n --ignore=vendor --extensions=php .",
+        "phpstan": "phpstan analyse ./src ./tests --level=0 --no-progress",
+        "generate-ssh-keys": "source ./vendor/keboola/db-writer-common/tests/generate-ssh-keys.sh",
+        "build": [
+            "@phpstan",
+            "@phpcs",
+            "@tests"
+        ],
+        "ci": [
+            "@composer validate --no-check-publish --no-check-all",
+            "@build"
+        ]
+    },
+    "config": {
+        "sort-packages": true,
+        "optimize-autoloader": true
     }
 }
-
