+++ phpunit.xml.dist
@@ -8,11 +8,7 @@
          processIsolation="false"
          stopOnFailure="true"
          syntaxCheck="false"
-         bootstrap="vendor/keboola/db-writer-common/bootstrap.php">
-
-    <php>
-        <const name="ROOT_PATH" value="./"/>
-    </php>
+         bootstrap="vendor/autoload.php">
 
     <testsuite name="MySQL Writer Test Suite">
         <directory>tests</directory>
