+++ tests/data/runAction/config.json
@@ -1,0 +1,69 @@
+{
+    "parameters": {
+        "db": {
+            "host": "mysql",
+            "port": "3306",
+            "database": "test",
+            "user": "root",
+            "#password": "rootpassword"
+        },
+        "tables": [
+            {
+                "tableId": "simple",
+                "dbName": "simple",
+                "export": true,
+                "incremental": false,
+                "primaryKey": ["id"],
+                "items": [
+                    {
+                        "name": "id",
+                        "dbName": "id",
+                        "type": "int",
+                        "size": null,
+                        "nullable": null,
+                        "default": null
+                    },
+                    {
+                        "name": "name",
+                        "dbName": "name",
+                        "type": "varchar",
+                        "size": 255,
+                        "nullable": null,
+                        "default": null
+                    },
+                    {
+                        "name": "glasses",
+                        "dbName": "glasses",
+                        "type": "varchar",
+                        "size": 255,
+                        "nullable": null,
+                        "default": null
+                    }
+                ]
+            },
+            {
+                "tableId": "special",
+                "dbName": "special",
+                "export": false,
+                "items": [
+                    {
+                        "name": "col1",
+                        "dbName": "col1",
+                        "type": "varchar",
+                        "size": 255
+                        "nullable":
+                        "default":
+                    },
+                    {
+                        "name": "col2",
+                        "dbName": "col2",
+                        "type": "varchar",
+                        "size": 255,
+                        "nullable": null,
+                        "default": null
+                    }
+                ]
+            }
+        ]
+    }
+}
